requirements_txt.txt
# Core ML and NLP libraries
torch>=1.9.0,<2.1.0
transformers>=4.20.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.21.0
scipy>=1.7.0

# PDF processing
PyMuPDF>=1.20.0
pdfminer.six>=20220524

# Language detection
langdetect>=1.0.9

# Web interfaces
gradio>=3.35.0
streamlit>=1.20.0

# Utilities
requests>=2.25.0
tqdm>=4.62.0
datasets>=2.0.0

# Optional: For better performance
# torch-audio  # Uncomment if needed
# accelerate   # Uncomment for faster model loading

# Development dependencies (optional)
jupyter>=1.0.0
ipython>=7.0.0

# System dependencies
psutil>=5.8.0